[FE] Fix: ProgressBar에서  Animate.ProgressBar 컴포넌트로 변경 및 수정
