+++ frontend/src/components/Animate/ProgressBar/ProgressBar.stories.tsx
@@ -3,10 +3,12 @@ import type { Meta, StoryObj } from "@storybook/react";
 import ProgressBar from "./ProgressBar";
 
 const meta = {
-  title: "Common/ProgressBar",
+  title: "Animate/ProgressBar",
   component: ProgressBar,
   argTypes: {
-    percent: { description: "공간을 채울 비율 ( 0 ~ 100 )" },
+    percent: { description: "" },
+    caption: { description: "" },
+    className: { description: "" },
   },
   tags: ["autodocs"],
   parameters: { layout: "padded" },
@@ -17,6 +19,7 @@ type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
   args: {
-    percent: 50,
+    percent: 60,
+    caption: "난이도(6)",
   },
 };
